Rename and move Alerts (non-full-width banners) as such
