+++ src/app/components/elements/alerts/RevisionWarningAlert.tsx
@@ -3,7 +3,7 @@ import {AppState, selectors, useAppSelector} from "../../../state";
 import {RenderNothing} from "../RenderNothing";
 import { Alert } from "reactstrap";
 
-export function RevisionWarningBanner() {
+export function RevisionWarningAlert() {
 
     const hideAttempts: boolean = useAppSelector(
         (state: AppState) => state?.userPreferences?.DISPLAY_SETTING?.HIDE_QUESTION_ATTEMPTS
